compute input mega blocks
Link: https://github.com/GeneAssembly/biosal/issues/260

+++ formats/fastq_input.h
@@ -7,6 +7,8 @@
 #include <input/buffered_reader.h>
 
 #include <stdio.h>
+#include <inttypes.h>
+#include <stdint.h>
 
 struct bsal_fastq_input {
     struct bsal_buffered_reader reader;
@@ -17,7 +19,7 @@ struct bsal_input_operations bsal_fastq_input_operations;
 
 void bsal_fastq_input_init(struct bsal_input *input);
 void bsal_fastq_input_destroy(struct bsal_input *input);
-int bsal_fastq_input_get_sequence(struct bsal_input *input,
+uint64_t bsal_fastq_input_get_sequence(struct bsal_input *input,
                 char *sequence);
 int bsal_fastq_input_detect(struct bsal_input *input);
